OGC Refactoring Part 9 - Move HTML converter service to package /ogc/services
