+++ server/src/main/java/de/ingrid/igeserver/ogc/exportCatalog/html/HtmlCatalogExporter.kt
@@ -21,7 +21,7 @@ package de.ingrid.igeserver.ogc.exportCatalog.html
 
 
 import com.fasterxml.jackson.databind.node.ObjectNode
-import de.ingrid.igeserver.ogc.OgcHtmlConverterService
+import de.ingrid.igeserver.ogc.services.OgcHtmlConverterService
 import de.ingrid.igeserver.ogc.exportCatalog.CatalogExportTypeInfo
 import de.ingrid.igeserver.ogc.exportCatalog.OgcCatalogExporter
 import de.ingrid.igeserver.persistence.postgresql.jpa.model.ige.Catalog
@@ -29,7 +29,6 @@ import de.ingrid.igeserver.services.CatalogService
 import de.ingrid.igeserver.services.DocumentCategory
 import de.ingrid.igeserver.services.DocumentService
 import org.keycloak.util.JsonSerialization.mapper
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Lazy
 import org.springframework.context.annotation.Profile
 import org.springframework.http.MediaType
