+++ server/src/main/java/de/ingrid/igeserver/ogc/services/OgcHtmlConverterService.kt
@@ -17,14 +17,13 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.ogc
+package de.ingrid.igeserver.ogc.services
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.JsonNodeType
 import com.fasterxml.jackson.databind.node.ObjectNode
 import de.ingrid.igeserver.configuration.GeneralProperties
 import de.ingrid.igeserver.ogc.model.Link
-import de.ingrid.igeserver.ogc.services.QueryMetadata
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Service
